README.md
When I saw the task I thought it could be done in a go and will be easy for me, But as I started the task problems kept on coming and it took me three days to finish the task.
So first i watched a tutorial on youtube on how to dual boot ubuntu 20.04, halfway through i realized that the method worked for systems with bios mode as legacy,
So i watched another video for bios mode UEFI,then I downloaded the ubuntu 20.04 and booted it onto a pen drive of 16 GB. when i booted it up for the first time, it failed and the maximum space of the pen drive became 3 MB, I was worried for a moment ,saw I watched a video on how to fix, Then I did as told and I got my pen drive as it is,
then I tried again, and I was successful in making a bootable pen drive with ubuntu. The app I used was rufus.io.
Then I started the booting process, during the installation process, a major problem occurred, my computer uses intel RST(Rapid Storage Technology)and it has to be turned of in order to continue,Ubuntu asked me to scan a QR code and follow the instructions it said. It took me 20 mins or so to understand what it said and even doing what it said didn't solve the problem. so I had to watch a video, and at last I was able to dual boot and run Ubuntu

But overall it was a great experience, I usually went to a friend or a computer store to do things related to computer, but I did this on my own with the help of google, which is the main takeaway from the task